Move options class into own header
